Fixed Upload Bugs
