+++ ui/src/components/pages/Summary.vue
@@ -112,11 +112,10 @@ export default {
       )
       return Object.fromEntries(basicPropsEntries)
     },
-
     summaryData: function() {
       if (this.fileData === {} || !this.fileData) return {}
       return {
-        filesize: this.bytesToSize(this.fileData.data.size),
+        filesize: this.fileData.data.size?this.bytesToSize(this.fileData.data.size):0,
         magic: this.fileData.data.magic,
         crc32: this.fileData.data.crc32,
         md5: this.fileData.data.md5,
